pyproject.toml
[project]
name = "cloudy_notes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "google-genai>=1.5.0",
    "libsql-experimental>=0.0.41",
    "mangum>=0.19.0",
    "ollama>=0.4.7",
    "pgvector>=0.3.6",
    "psycopg[binary,pool]>=3.2.5",
    "urllib3>=2.3.0",
    "uvicorn>=0.34.0",
]
